requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
click==8.2.1
discord.py==2.5.2
Flask==3.1.1
frozenlist==1.6.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
propcache==0.3.1
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
Werkzeug==3.1.3
yarl==1.20.0
